added AssertionException, and moved Assert and AssertionException to a subfolder in the util directory
