+++ lib/Wikia/src/Util/ErrorChecking/AssertionException.php
@@ -1,0 +1,15 @@
+<?php
+/**
+ * AssertionException
+ *
+ * Exception denoting a failed error check (such as from Assert)
+ *
+ * @author Nelson Monterroso <nelson@wikia-inc.com>
+ */
+
+namespace Wikia\Util\ErrorChecking;
+
+
+class AssertionException extends \Exception {
+
+}
